implement: addiu, lw, lh, lhu, lb

+++ 103062224_01/simulator/simulator.cpp
@@ -8,14 +8,18 @@
 #include "./regfile.h"
 #include "./memory.h"
 
+FILE *snapshot;
+FILE *error_dump;
+
 Memory* memory;
 InstructionMemery* instructions;
 Registers* regs;
 unsigned int PC;
+int cycle;
 
 std::vector<unsigned int>* readImage(FILE *);
 void readInput_initialize(void);
-void print_snapshot(FILE *, int);
+void print_snapshot(void);
 int execute(void);
 void destroy_all(void);
 
@@ -30,16 +34,23 @@ int main(int argc, char const *argv[])
 	//for(int i=0 ; i<regs->size() ; i++) printf("%x ", regs->at(i));
 	//printf("\n") ;
 
-	int cycle = 0;
-	FILE *snapshot = fopen("snapshot.rpt", "w");
+	cycle = 0;
+	snapshot = fopen("snapshot.rpt", "w");
 	if(snapshot==NULL){
 		fputs("snapshot write error", stderr);
-		exit(2);
+		exit(666);
 	}
-	print_snapshot(snapshot, cycle);
+	error_dump = fopen("error_dump.rpt", "w");
+	if(error_dump==NULL){
+		fputs("error_dump write error", stderr);
+		exit(666);
+	}
+
+
+	print_snapshot();
 	while (execute() != -1){
 		cycle++;
-		print_snapshot(snapshot, cycle);
+		print_snapshot();
 	}
 
 	destroy_all();
@@ -85,6 +96,7 @@ void readInput_initialize(void)
 
 int execute(void)
 {
+	bool toReturn = false;
 	unsigned int inst = instructions->at(PC/4);
 	PC += 4;
 	//printf("inst==%x  ", inst);
@@ -140,14 +152,122 @@ int execute(void)
 			default: break;
 		}
 	} else {
-		unsigned int immediate = inst & 0xffff;
+		unsigned short immediate = inst & 0xffff;
 		unsigned int address = inst & 0x3ffffff;
+		unsigned int location;
+		signed short halfLoaded;
+		signed char byteLoaded;
 		switch(opcode)
 		{
+
+			//printf("immediate==%d\n", immediate);
+			//--------------------------- I type start -----------------------------//
 			case 0x08: 	// addi
-				//printf("immediate==%d\n", immediate);
-				regs->at(rt) = regs->at(rs) + immediate;
+				regs->at(rt) = regs->at(rs) + ((signed short)immediate);
 				break;
+
+			case 0x09:	// addiu
+				regs->at(rt) = regs->at(rs) + ((unsigned short)immediate);
+				break;
+
+			case 0x23:	//lw
+				location = (rs + ((signed short)immediate) );
+				if ( location >1020 ) {
+					fprintf(error_dump, "In cycle %d: Address Overflow\n", cycle);
+					toReturn = true;
+				}
+				if( location % 4 != 0 ){
+					fprintf(error_dump, "In cycle %d: Misalignment Error\n", cycle);
+					toReturn = true;
+				}
+				if(toReturn) return -1;
+
+				regs->at(rt) = memory->at( location/4 );
+				break;
+
+			case 0x21:	//lh
+				location = (rs + ((signed short)immediate) );
+				if( location > 1022) {
+					fprintf(error_dump, "In cycle %d: Address Overflow\n", cycle);
+					toReturn = true;
+				}
+				if( location % 2 != 0){
+					fprintf(error_dump, "In cycle %d: Misalignment Error\n", cycle);
+					toReturn = true;
+				}
+				if(toReturn) return -1;
+
+				if(location%4==0) halfLoaded = (signed short) ( (regs->at(location/4)) >> 16);
+				else if(location%2==0) halfLoaded = (signed short) ( (regs->at(location)) & 0x0000ffff );
+				regs->at(rt) = (signed short)halfLoaded;		// <-------- this line is very important!!!
+				break;
+
+			case 0x25:	//lhu 
+				location = (rs + ((signed short)immediate) );
+				if( location > 1022) {
+					fprintf(error_dump, "In cycle %d: Address Overflow\n", cycle);
+					toReturn = true;
+				}
+				if( location % 2 != 0){
+					fprintf(error_dump, "In cycle %d: Misalignment Error\n", cycle);
+					toReturn = true;
+				}
+				if(toReturn) return -1;
+
+				if(location%4==0) halfLoaded = (unsigned short) ( (regs->at(location/4)) >> 16);
+				else if(location%2==0) halfLoaded = (unsigned short) ( (regs->at(location)) & 0x0000ffff );
+				regs->at(rt) = (unsigned short)halfLoaded;
+				break;
+
+			case 0x20:	//lb 
+				location = (rs + ((signed short)immediate) );
+				if( location > 1023) {
+					fprintf(error_dump, "In cycle %d: Address Overflow\n", cycle);
+					toReturn = true;
+				}
+				if(toReturn) return -1;
+
+				byteLoaded = (signed char) (( regs->at(location/4) ) >> (	(location%4==0) ? 24 :
+																																	(location%4==1) ? 16 :
+																																	(location%4==2) ? 8  :
+																																	(location%4==3) ? 0 : 0) ) & 0x000000ff; 
+
+				regs->at(rt) = (signed char)byteLoaded;
+				break;
+			case 0x24:	//lbu 
+				break;
+			case 0x2B:	//sw 
+				break;
+			case 0x29:	//sh 
+				break;
+			case 0x28:	//sb 
+				break;
+			case 0x0F:	//lui 
+				break;
+			case 0x0C:	//andi 
+				break;
+			case 0x0D:	//ori 
+				break;
+			case 0x0E:	//nori 
+				break;
+			case 0x0A:	//slti 
+				break;
+			case 0x04:	//beg 
+				break;
+			case 0x05:	//bne 
+				break;
+			case 0x07:	//bgtz 
+				break;
+			//--------------------------- I type end -----------------------------//
+
+
+			//--------------------------- J type start -----------------------------//
+			/*case 0x:	// 
+				break;*/
+			//--------------------------- J type end -----------------------------//
+
+
+			
 			case 0x3f:	// halt
 				return -1;
 				break;
@@ -160,7 +280,7 @@ int execute(void)
 
 
 
-void print_snapshot(FILE *snapshot, int cycle)
+void print_snapshot(void)
 {
 	fprintf(snapshot , "cycle %d\n", cycle);
 	for(unsigned int i=0 ; i<regs->size() ; i++){
@@ -178,6 +298,9 @@ void destroy_all(void)
 	delete memory;
 	delete instructions;
 	delete regs;
+
+	fclose(snapshot);
+	fclose(error_dump);
 }
